bugfix suggested tasks
